refactor: extract engine.rs inline tests to tests/engine_tests.rs
Move public-API tests (resize_calc, security, apply_ops, optimize_ops,
encode, decode_mozjpeg, fast_resize) to tests/engine_tests.rs.
Keep ICC and EncodeTask/FirewallConfig tests inline since they depend
on pub(crate) internals.
